+ move FillType.h to App folder
